Tic-Tac-Toe and Battle Arena changes
Tic-Tac-Toe will print "Draw!" to console when there are no more available spaces on the board.
Battle Arena has minimum and maximum damage values, and randomly deal damage in between the two values.
It will print to console what damage has been dealt to the target, as well as the current state of the target's health after the fact.